perf(train): fix memory fragmentation, GDR learning, RoPE dtype, sharded checkpoints
Memory: remove torch.cuda.empty_cache() calls causing VRAM fragmentation.
Config: add num_workers:0, pin_memory:false to training section.

GDR: guard auxiliary_output with isinstance(gdr_output, dict) — prevents
L2 regularization on hidden states which kills learning.

RoPE: create cos/sin directly in target dtype instead of float32->to(dtype).
Eliminates 149ms aten::to per step. Speed +40% (6873 -> 9594 tok/s).

Checkpoint: load_checkpoint supports sharded checkpoints (model.pt + ema.pt
+ meta.pt) from save_training_checkpoint with MoE.

Tests: 214 passed, 1 skipped.

Constraint: Windows RTX 3070, no MSVC for torch.compile.
Rejected: manual_fp16 (loss explodes), bf16 (slower than fp16+autocast).
Not-tested: 8000-step full run (interrupted at step 1050).

+++ src/hagi/model/transformer.py
@@ -62,11 +62,11 @@ class RMSNorm(nn.Module):
 
 
 def build_rope_cache(seq_len: int, head_dim: int, theta: float, device, dtype):
-    inv_freq = 1.0 / (theta ** (torch.arange(0, head_dim, 2, device=device).float() / head_dim))
-    t = torch.arange(seq_len, device=device).float()
+    inv_freq = 1.0 / (theta ** (torch.arange(0, head_dim, 2, device=device, dtype=dtype) / head_dim))
+    t = torch.arange(seq_len, device=device, dtype=dtype)
     freqs = torch.outer(t, inv_freq)
-    cos = freqs.cos().to(dtype)
-    sin = freqs.sin().to(dtype)
+    cos = freqs.cos()
+    sin = freqs.sin()
     return cos, sin
